requirements.txt
langchain==0.3.21
langchain_community==0.3.20
langchain_core==0.3.47
langchain_huggingface==0.1.2
langchain_ollama==0.3.0
ollama==0.4.7
pipreqs==0.4.13
python-dotenv==1.0.1
streamlit==1.43.2
torch==2.6.0+cu126
tqdm==4.67.1
transformers==4.49.0
